Refactor source files from './src/ to ../src
